🧰💭🎛️👼: remove `.viewModel` accesses
