+++ lively.ide/world-commands.js
@@ -583,7 +583,7 @@ const commands = [
       halo.alignWithTarget();
       once(responsiveHalo, 'close', halo.responsiveHalo(), 'stop');
       once(halo, 'remove', responsiveHalo, 'close');
-      world.propertiesPanel?.get('responsive control').viewModel.activate();
+      world.propertiesPanel?.get('responsive control').activate();
 
       return responsiveHalo;
     }
@@ -864,10 +864,10 @@ const commands = [
         const { ComponentBrowser } = await System.import('lively.ide/studio/component-browser.cp.js');
         const componentBrowser = world._componentBrowser || (world._componentBrowser = part(ComponentBrowser, { name: 'lively component browser' }));
         li.remove();
-        $world.get('lively top bar')?.colorTopbarButton($world.get('lively top bar').viewModel.ui.openComponentBrowser, true);
+        $world.get('lively top bar')?.colorTopbarButton($world.get('lively top bar').get('open component browser'), true);
         componentBrowser.openInWindow({ minimizable: false, title: 'Browse Components' });
         const loadedComponent = await componentBrowser.activate();
-        $world.get('lively top bar')?.colorTopbarButton($world.get('lively top bar').viewModel.ui.openComponentBrowser, false);
+        $world.get('lively top bar')?.colorTopbarButton($world.get('lively top bar').get('open component browser'), false);
         if (loadedComponent && !loadedComponent.world()) { loadedComponent.openInWorld(); }
       })();
     }
